requirements.txt
fastapi
uvicorn[standard]
torch
transformers>=4.40.0
accelerate
pillow
requests
